# Conflicts:
#	ecomerge/lib/database/database_helper.dart
#	ecomerge/lib/database/services/user_service.dart

+++ Server/src/main/java/demo/com/example/testserver/product/model/Product.java
@@ -26,7 +26,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id; // Changed from Integer to Long
 
     @Column(name = "ten_san_pham", nullable = false)
     private String name;
@@ -92,11 +92,11 @@ public class Product {
     public Product() {}
 
     // Getters and Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
